+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 23.1.4
 - TRIO114 no longer triggers on posonly args named "task_status"
 - TRIO116 will now match on any attribute parameter named `.inf`, not just `math.inf`.
+- TRIO900 now only checks `@asynccontextmanager`, not other decorators passed with --no-checkpoint-warning-decorators.
 
 ## 23.1.3
 - Add TRIO240: usage of `os.path` in async function.